Read chapter 1. ANSI C

+++ Fahrenheit-celsius.c
@@ -8,7 +8,7 @@
  * print fahrenheit -> celsius
  */
 int fahrenheitcelsius() {
-    int fahr, celsius;
+    float fahr, celsius;
     int lower, upper, step;
 
     lower = 0;
@@ -18,7 +18,7 @@ int fahrenheitcelsius() {
     fahr = lower;
     while (fahr <= upper) {
         celsius = 5 * (fahr-32) / 9;
-        printf("%d\t%d\n", fahr, celsius);
+        printf("%3.0f\t%6.1f\n", fahr, celsius);
         fahr = fahr + step;
     }
     return 0;
